test(picker): migrate is-open
